Python/requirements.txt
azure-functions
azure-keyvault-secrets
azure.identity
adal